pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "ghostx"
version = "1.0.0"
description = "A modern, secure, and feature-rich email management system"
authors = ["xtial <xtial@github.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/xtial/GhostX"
repository = "https://github.com/xtial/GhostX"
documentation = "https://github.com/xtial/GhostX/docs"
keywords = ["email", "management", "security", "flask", "web"]
packages = [
    { include = "src" }
]
package-mode = false
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Web Environment",
    "Framework :: Flask",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
    "Topic :: Communications :: Email",
    "Topic :: Security",
]

[tool.poetry.dependencies]
python = "^3.13"
flask = "^3.0.0"
flask-sqlalchemy = "^3.1.1"
flask-login = "^0.6.3"
flask-wtf = "^1.2.1"
flask-migrate = "^4.0.5"
flask-session = "^0.5.0"
flask-limiter = "^3.5.0"
flask-talisman = "^1.1.0"
flask-compress = "^1.14.0"
flask-cors = "^4.0.0"
sqlalchemy = "^2.0.23"
wtforms = "^3.1.1"
email-validator = "^2.1.0"
python-dotenv = "^1.0.0"
bcrypt = "^4.0.1"
pyjwt = "^2.8.0"
requests = "^2.31.0"
gunicorn = "^23.0.0"
psycopg2-binary = "^2.9.9"
redis = "^5.0.1"
pillow = "^10.1.0"
pydantic = "^2.5.2"
click = "^8.1.7"
passlib = "^1.7.4"
beautifulsoup4 = "^4.12.2"
psutil = "^5.9.8"
typing-extensions = "^4.9.0"
werkzeug = "^3.0.1"
jinja2 = "^3.1.3"
itsdangerous = "^2.1.2"
markupsafe = "^2.1.5"
blinker = "^1.7.0"
packaging = "^23.2"
waitress = "^3.0.0"
prometheus-client = "^0.19.0"
sentry-sdk = "^1.39.1"
statsd = "^4.0.1"
flask-caching = ">=2.1.0"
bleach = "^6.2.0"

[tool.poetry.group.dev.dependencies]
black = "^23.9.0"
flake8 = "^6.1.0"
mypy = "^1.5.0"
bandit = "^1.7.0"
safety = "^2.3.0"
pre-commit = "^3.4.0"

[tool.poetry.scripts]
ghostx = "setup:run_server"

[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'

[tool.mypy]
python_version = "3.10"
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
warn_unreachable = true
strict_optional = true 